+++ apps/sr/bsa_cli/commands/ExtractAll.cpp
@@ -1,7 +1,7 @@
 /*
  -------------------------------------------------------------------------------
     This file is part of the Skyrim Tools Project.
-    Copyright (C) 2021, 2022, 2023  Dirk Stolle
+    Copyright (C) 2021, 2022, 2023, 2025  Dirk Stolle
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -103,9 +103,10 @@ int ExtractAll::run()
   const auto& header = bsa.getHeader();
   if (!header.hasNamesForDirectories() || !header.hasNamesForFiles())
   {
-    std::cout << "Info: The file " << bsaFileName << " does not contain "
-              << "information about its directory names and file names.\n";
-    return 0;
+    std::cerr << "Error: The file " << bsaFileName << " does not contain "
+              << "information about its directory names and file names. The "
+              << "program cannot handle such an archive.\n";
+    return SRTP::rcDataError;
   }
   if (!bsa.grabAllStructureData())
     return SRTP::rcFileError;