+++ apps/sr/bsa_cli/main.cpp
@@ -34,7 +34,7 @@
 
 void showVersion()
 {
-  std::cout << "BSA Command Line Tool for Skyrim, version 0.11.1, 2024-05-05\n"
+  std::cout << "BSA Command Line Tool for Skyrim, version 0.12.0, 2025-05-14\n"
             << "\n"
             << "Copyright (C) 2021 - 2025  Dirk Stolle\n"
             << "License GPLv3+: GNU GPL version 3 or later <https://gnu.org/licenses/gpl.html>\n"